Create PatchMeta component
