+++ client/src/Pages/Admin/components/Dropdown.css
@@ -46,7 +46,9 @@
     border: solid 1px #e4e4e4;
     cursor: pointer;
     width: 100%;
+    max-height: 100px;
     user-select: none;
+    overflow: hidden;
 }
 
 .Dropdown .results .item:hover {
@@ -72,7 +74,7 @@
 
 .Dropdown .results .item .col.content .meta {
     color: rgba(0, 0, 0, 0.6);
-    margin: 10px 0 0 0;
+    /* margin: 10px 0 0 0; */
     overflow: hidden;
 }
