+++ client/src/Pages/PatchDetailed/components/PatchMeta.css
@@ -1,0 +1,19 @@
+.PatchMeta .meta {
+    color: rgba(0, 0, 0, 0.6);
+}
+
+.PatchMeta .meta i.fa-circle {
+    font-size: 0.5em;
+    position: relative;
+    top: -2px;
+    padding: 0 5px 0 5px;
+    color: rgba(0, 0, 0, 0.3);
+}
+
+.PatchMeta .meta #cName:not(:last-child)::after {
+    content: ", ";
+}
+
+.PatchMeta .meta #cName:first-child::before {
+    content: " ";
+}
