+++ client/src/Pages/PatchDetailed/components/PatchMeta.js
@@ -1,0 +1,46 @@
+import React from 'react'
+import moment from 'moment'
+import './PatchMeta.css'
+import { PRICE_TYPES } from '../../../config/constants'
+
+
+const PatchMeta = ({data}) => {
+    const priceDisplay = _ => {
+        const price = data.price
+        if (price.type === PRICE_TYPES.SET_PRICE) {
+            return price.value + " SEK"
+        } else return price.type
+    }
+
+    return (
+        <div className="PatchMeta">
+            <div className="meta">
+                <span title="Första försäljningsdatum">
+                    <i className="far fa-clock"></i> {data.date ? moment(data.date).format("DD MMM YYYY") : "Okänt"}
+                </span>
+                <i className="fas fa-circle"></i>
+                <span title="Uppladdningsdatum"><i className="fas fa-cloud-upload-alt"></i> {moment(data.createdAt).format("DD MMM YYYY")}</span>
+                <i className="fas fa-circle"></i>
+                <span title="Pris">
+                    <i className="fas fa-dollar-sign"></i> {priceDisplay()}
+                </span>
+                <i className="fas fa-circle"></i>
+                <span title="Lagerstatus (till salu)">
+                    <i className="fas fa-box-open"></i> {data.inStock === true ? "I lager" : "Ej i lager"}
+                </span>
+                <i className="fas fa-circle"></i>
+                <span title="Antal producerade (stämmer ej för märken äldre än Damm)">
+                    <i className="fas fa-hashtag"></i> {data.produced} st
+                </span>
+            </div>
+            {data.creators.length !== 0 && 
+                <div className="meta">
+                        Skapad av:
+                        {data.creators.map(c => <span id="cName">{c.name}</span>)}
+                </div>
+            }
+        </div>
+    )
+}
+
+export default PatchMeta
